Operation unfuck the dialogs
The dialogs in this app were just not surviving a rotation nor were they encouraging
users to use the dialog buttons. This fixes that. All dialogs now survive a rotation
and the user has to press 'OK' to exit out of the dialog. Some people don't like this
but this behavior but this is my personal preference.

- Get rid of the check on first boot that checks for Lineage or Omni. This change
was not only useless but also caused the app to lag a bit on first boot.

- Removed unnecessary menu items (feedback, fresh and info). Credits/About was retained
so there's no need for a 'info' menu item.

- Removed 'report missing device' option. All our devices will be added. No need to have
this option. Also TWRP filter option was removed. We don't handle TWRP on our gerrit.

- Oranged strings because it was a hot mess before. Now we know what we're working with.

- Some more layout/theme changes